requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
psycopg2-binary
python-dotenv
crewai
openai
google-cloud-aiplatform
Flask-Migrate
Flask-Cors
Flask>=2.0
Flask-Migrate
psycopg2-binary
crewAI
python-dotenv
SQLAlchemy
litellm
langchain
